PRSD-804: Commonise always true validator
