+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/forms/LandlordRegistrationJourneyTests.kt
@@ -13,12 +13,11 @@ import org.mockito.kotlin.whenever
 import org.springframework.security.core.Authentication
 import org.springframework.security.core.context.SecurityContext
 import org.springframework.security.core.context.SecurityContextHolder
-import org.springframework.validation.Errors
-import org.springframework.validation.Validator
 import uk.gov.communities.prsdb.webapp.database.entity.LocalAuthority
 import uk.gov.communities.prsdb.webapp.forms.journeys.LandlordRegistrationJourney
 import uk.gov.communities.prsdb.webapp.forms.journeys.PageData
 import uk.gov.communities.prsdb.webapp.forms.steps.LandlordRegistrationStepId
+import uk.gov.communities.prsdb.webapp.helpers.AlwaysTrueValidator
 import uk.gov.communities.prsdb.webapp.mockObjects.JourneyDataBuilder
 import uk.gov.communities.prsdb.webapp.mockObjects.MockLandlordData
 import uk.gov.communities.prsdb.webapp.models.viewModels.emailModels.LandlordRegistrationConfirmationEmail
@@ -50,15 +49,6 @@ class LandlordRegistrationJourneyTests {
     lateinit var urlProvider: AbsoluteUrlProvider
 
     val alwaysTrueValidator: AlwaysTrueValidator = AlwaysTrueValidator()
-
-    class AlwaysTrueValidator : Validator {
-        override fun supports(clazz: Class<*>): Boolean = true
-
-        override fun validate(
-            target: Any,
-            errors: Errors,
-        ) {}
-    }
 
     @BeforeEach
     fun setup() {
