+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/helpers/AlwaysTrueValidator.kt
@@ -1,0 +1,13 @@
+package uk.gov.communities.prsdb.webapp.helpers
+
+import org.springframework.validation.Errors
+import org.springframework.validation.Validator
+
+class AlwaysTrueValidator : Validator {
+    override fun supports(clazz: Class<*>): Boolean = true
+
+    override fun validate(
+        target: Any,
+        errors: Errors,
+    ) {}
+}
